+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## 1.4.4 - 18.04.2020
+
+* add `debug` (optional) parameter in `initialize()` method that supports disable logging to console
+
+## 1.4.3 - 09.04.2020
+
+* iOS: fix bug on `remove` method
+
+## 1.4.2 - 02.04.2020
+
+* add `timeCreated` in `DownloadTask` model
+* iOS: fix bug MissingPluginException
+
 ## 1.4.1 - 30.01.2020
 
 * Android: fix bug `ensureInitializationComplete must be called after startInitialization`